create reusable schema, update testcases
