--- test/metadata/testCreateCustomObject.json
@@ -1,75 +1,0 @@
-{
-    "xmlName": "metadata",
-    "attributes": [
-        {
-            "key": "xsi:type",
-            "value": "CustomObject"
-        }
-    ],
-    "children": [
-        {
-            "xmlName": "fullName",
-            "attributes": null,
-            "children": [
-                "TestObject15__c"
-            ],
-            "selfClosing": false
-        },
-        {
-            "xmlName": "label",
-            "attributes": null,
-            "children": [
-                "Test Object 15"
-            ],
-            "selfClosing": false
-        },
-        {
-            "xmlName": "pluralLabel",
-            "attributes": null,
-            "children": [
-                "Test Objects 15"
-            ],
-            "selfClosing": false
-        },
-        {
-            "xmlName": "nameField",
-            "attributes": null,
-            "children": [
-                {
-                    "xmlName": "type",
-                    "attributes": null,
-                    "children": [
-                        "Text"
-                    ],
-                    "selfClosing": false
-                },
-                {
-                    "xmlName": "label",
-                    "attributes": null,
-                    "children": [
-                        "Test Object Name"
-                    ],
-                    "selfClosing": false
-                }
-            ],
-            "selfClosing": false
-        },
-        {
-            "xmlName": "deploymentStatus",
-            "attributes": null,
-            "children": [
-                "Deployed"
-            ],
-            "selfClosing": false
-        },
-        {
-            "xmlName": "sharingModel",
-            "attributes": null,
-            "children": [
-                "ReadWrite"
-            ],
-            "selfClosing": false
-        }
-    ],
-    "selfClosing": false
-}
